refactor(Lua): separate dialog and manager implementation
